Update_final_recog_face.py
This script is the main application. It uses the camera to detect faces, identifies them with the trained trainer.yml model, and draws their name on the screen. It then automatically logs their "ENTRY" or "EXIT" time into the attendance_log.csv file. And also mark the detected face by their name in data set or unknow in not present in the data set.